requirements.txt
# Stage 4 MVP Pipeline Dependencies
# Python 3.9+

# Core
openai>=1.0.0
pydantic>=2.0.0
jsonschema>=4.0.0

# Async HTTP
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
ipykernel>=6.0.0
jupyter>=1.0.0
